requirements.txt
openai==1.12.0
psycopg2-binary==2.9.9
argparse==1.4.0
python-dotenv==1.0.1
aider
tiktoken